+++ app/models/account_authz/current.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module AccountAuthz
+  class Current < ActiveSupport::CurrentAttributes
+    attribute :account_id
+  end
+end