+++ app/models/citizen/assignment.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-module Citizen
-  class Assignment < ApplicationRecord
-    belongs_to :member, polymorphic: true
-    belongs_to :role, class_name: "Citizen::Role"
-  end
-end