+++ test/dummy/app/models/member.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
 class Member < ApplicationRecord
-  include Citizen::Member
+  include AccountAuthz::Member
 end