+++ test/dummy/config/initializers/citizen.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-Rails.application.config.to_prepare do
-  Citizen.members_source = MemberDirectory
-end